backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
CodeConvert==3.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.16
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==23.1
django-models-ext==1.1.11
django-short-url==1.1.8
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.15.2
djoser==2.1.0
drf-extra-fields==3.7.0
excel-base==1.0.4
filetype==1.2.0
furl==2.1.3
idna==3.10
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
orderedmultidict==1.0.1
pillow==11.0.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
xlwt==1.3.0
